jktech/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.0.335
langchain-community==0.0.13
sentence-transformers==2.2.2
numpy==1.26.1
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
asyncpg==0.28.0
scikit-learn==1.3.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
alembic==1.12.1
